+++ src/camera.h
@@ -10,19 +10,19 @@ class Camera
 public:
     const Vector& GetPosition() const { return m_Position; }
     void SetPosition(const Vector& position) { m_Position = position; }
-    float GetYaw() const { return m_Yaw; }
-    void SetYaw(float Yaw) { m_Yaw = Yaw; }
-    float GetPitch() const { return m_Pitch; }
-    void SetPitch(float Pitch) { m_Pitch = Pitch; }
-    float GetRoll() const { return m_Roll; }
-    void SetRoll(float Roll) { m_Roll = Roll; }
-    float GetAspectRatio() const { return m_AspectRatio; }
-    void SetAspectRatio(float AspectRatio) { m_AspectRatio = AspectRatio; }
-    float GetFOV() const { return m_FOV; }
-    void SetFOV(float FOV) { m_FOV = FOV; }
+    double GetYaw() const { return m_Yaw; }
+    void SetYaw(double Yaw) { m_Yaw = Yaw; }
+    double GetPitch() const { return m_Pitch; }
+    void SetPitch(double Pitch) { m_Pitch = Pitch; }
+    double GetRoll() const { return m_Roll; }
+    void SetRoll(double Roll) { m_Roll = Roll; }
+    double GetAspectRatio() const { return m_AspectRatio; }
+    void SetAspectRatio(double AspectRatio) { m_AspectRatio = AspectRatio; }
+    double GetFOV() const { return m_FOV; }
+    void SetFOV(double FOV) { m_FOV = FOV; }
 
     void FrameBegin();
-    Ray GetScreenRay(float x, float y) const;
+    Ray GetScreenRay(double x, double y) const;
 
 private:
     Vector m_TopLeft;
@@ -32,12 +32,12 @@ private:
     Vector m_Position;
     Matrix m_Rotation;
 
-    float m_Yaw;
-    float m_Pitch;
-    float m_Roll;
+    double m_Yaw;
+    double m_Pitch;
+    double m_Roll;
 
-    float m_AspectRatio;
-    float m_FOV;
+    double m_AspectRatio;
+    double m_FOV;
 };
 
 #endif //RAYTRACING_CAMERA_H
